Add full meta-analytic CDat combination to Combiner   Equivalent to weighted mean; should be used with z-score normalization   Seems to perform almost identically to unweighted and is ~4-6x slower Add the long-awaited ability to do non-discretized comparisons (poorly) in MIer Add genewise context inference (multithreaded) to Counter Add CDat::Clear function to initialize non-NaN values Fix some gcc floating point compilation glitches in CPCL and Funcifier

+++ src/pcl.cpp
@@ -881,10 +881,10 @@ void CPCL::Normalize( ENormalize eNormalize ) {
 					for( j = 0; j < GetGenes( ); ++j )
 						if( !CMeta::IsNaN( d = Get( j, i ) ) ) {
 							if( eNormalize == ENormalizeColumnFraction )
-								d /= dAve;
+								d /= (float)dAve;
 							else
-								d = ( d - dAve ) / dStd;
-							Set( j, i, (float)d ); } } }
+								d = (float)( ( d - dAve ) / dStd );
+							Set( j, i, d ); } } }
 			break;
 
 		case ENormalizeMinMax:
